ReadOnlySpan<byte> over static readonly byte[]
https://vcsjones.dev/csharp-readonly-span-bytes-static/
no need to ever load them on the heap

+++ PKHeX.Core/Legality/Restrictions/Vivillon3DS.cs
@@ -273,7 +273,7 @@ public static class Vivillon3DS
     /// </summary>
     /// <remarks>Not to be confused with <see cref="Region3DSIndex"/>.</remarks>
     [Flags]
-    private enum Region3DSFlags : ushort
+    private enum Region3DSFlags : byte
     {
         None,
         Japan = 1,
